Clean up Drosera so that it will build 64-bit.
Reviewed by Darin Adler.

* Drosera/mac/DebuggerApplication.mm:
(-[DebuggerApplication numberOfRowsInTableView:]): Use NSInteger rather than int.
(-[DebuggerApplication tableView:objectValueForTableColumn:row:]): Ditto.
* Drosera/mac/Drosera.xcodeproj/project.pbxproj: Use the default value for VALID_ARCHS.


Canonical link: https://commits.webkit.org/26190@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@32994 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKitTools/Drosera/mac/DebuggerApplication.mm
@@ -33,6 +33,7 @@
 #import "ServerConnection.h"
 
 #import <WebKit/WebCoreStatistics.h>
+#import <WebKit/WebTypesInternal.h>
 
 @implementation DebuggerApplication
 - (void)awakeFromNib
@@ -117,17 +118,17 @@
 #pragma mark -
 #pragma mark Table View Delegate
 
-- (int)numberOfRowsInTableView:(NSTableView *)tableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tableView
 {
     return [knownServerNames count];
 }
 
-- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(int)row
+- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {
     return @"";
 }
 
-- (void)tableView:(NSTableView *)tableView willDisplayCell:(id)cell forTableColumn:(NSTableColumn *)tableColumn row:(int)row
+- (void)tableView:(NSTableView *)tableView willDisplayCell:(id)cell forTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {
     NSString *key = [[knownServerNames allKeys] objectAtIndex:row];
     NSMutableDictionary *info = [knownServerNames objectForKey:key];
